+++ packages/agent/src/agent/http/index.ts
@@ -42,6 +42,7 @@ import { Ed25519PublicKey } from '../../public_key';
 import { decodeTime } from '../../utils/leb';
 import { ObservableLog } from '../../observable';
 import { BackoffStrategy, BackoffStrategyFactory, ExponentialBackoff } from '../../polling/backoff';
+import { DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS } from '../../constants';
 export * from './transforms';
 export { Nonce, makeNonce } from './types';
 
@@ -53,9 +54,6 @@ export enum RequestStatusResponseStatus {
   Unknown = 'unknown',
   Done = 'done',
 }
-
-// Default delta for ingress expiry is 5 minutes.
-const DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS = 5 * 60 * 1000;
 
 // Root public key for the IC, encoded as hex
 export const IC_ROOT_KEY =
@@ -785,7 +783,6 @@ export class HttpAgent implements Agent {
 
       const requestId = await requestIdOf(request);
 
-      // TODO: remove this any. This can be a Signed or UnSigned request.
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       let transformedRequest: HttpAgentRequest = await this._transform({
         request: {
@@ -952,9 +949,8 @@ export class HttpAgent implements Agent {
     }
     const sender = id?.getPrincipal() || Principal.anonymous();
 
-    // TODO: remove this any. This can be a Signed or UnSigned request.
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const transformedRequest: any = await this._transform({
+    const transformedRequest = await this._transform({
       request: {
         method: 'POST',
         headers: {
@@ -985,7 +981,14 @@ export class HttpAgent implements Agent {
     const canister = typeof canisterId === 'string' ? Principal.fromText(canisterId) : canisterId;
 
     const transformedRequest = request ?? (await this.createReadStateRequest(fields, identity));
-    const body = cbor.encode(transformedRequest.body);
+
+    // With read_state, we should always use a fresh expiry, even beyond the point where the initial request would have expired
+    const bodyWithAdjustedExpiry = {
+      ...transformedRequest.body,
+      ingress_expiry: new Expiry(DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS),
+    };
+
+    const body = cbor.encode(bodyWithAdjustedExpiry);
 
     this.log.print(
       `fetching "/api/v2/canister/${canister}/read_state" with request:`,