+++ packages/agent/src/constants.ts
@@ -1,0 +1,2 @@
+// Default delta for ingress expiry is 5 minutes.
+export const DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS = 5 * 60 * 1000;